Fixed error on win11

+++ text_rendering.h
@@ -13,10 +13,10 @@
 #include <glad/glad.h>
 
 struct Character {
-    unsigned int TextureID;  // ID handle of the glyph texture
-    glm::ivec2   Size;       // Size of glyph
-    glm::ivec2   Bearing;    // Offset from baseline to left/top of glyph
-    unsigned int Advance;    // Offset to advance to next glyph
+    unsigned int TextureID; // ID handle of the glyph texture
+    glm::ivec2   Size;      // Size of glyph
+    glm::ivec2   Bearing;   // Offset from baseline to left/top of glyph
+    long int Advance;       // Offset to advance to next glyph
 };
 
 std::map<char, Character> Characters;
